backend/pyproject.toml
[project]
name = "job-scheduler-backend"
version = "0.1.0"
description = "FastAPI backend for job scheduling, queueing, tracking, snapshots, and logging."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "pydantic-settings>=2.6.0,<3.0.0",
  "sqlalchemy>=2.0.36,<3.0.0",
  "psycopg[binary]>=3.2.0,<4.0.0",
  "uvicorn[standard]>=0.32.0,<1.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]
